Add files via upload
In a nutshell, the purpose of this script is to validate satellite data as a quality control measure. This project was done for the federal NOAA Center for Weather and Climate Prediction. In this script, Satellite soil moisture data was opened in a binary file and appeared in a matrix representing a map. Each cell in the matrix was a geographical coordinate, which was to be matched with its respective weather station located on the ground. After matching all sites with its geographical coordinate, statistical analysis was performed to assess the correlation between values.  An example of the visual product put out by the code is attached, where satellite data covering Hodges, Alabama was matched in-situ weather station and compared.